some small re-org
